combine comparators
